// services/database.types.ts
// This file is auto-generated by Supabase.
// It is recommended to run "npx supabase gen types typescript --project-id lsfjxtrqggogkzefudry > services/database.types.ts" to keep it in sync.

export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export interface Database {
  public: {
    Tables: {
      lead: {
        Row: {
          id: number
          created_at: string
          post_title: string | null
          post_url: string | null
          author_name: string | null
          author_url: string | null
          author_description: string | null
        }
        Insert: {
          id?: number
          created_at?: string
          post_title?: string | null
          post_url?: string | null
          author_name?: string | null
          author_url?: string | null
          author_description?: string | null
        }
        Update: {
          id?: number
          created_at?: string
          post_title?: string | null
          post_url?: string | null
          author_name?: string | null
          author_url?: string | null
          author_description?: string | null
        }
      }
      my_lead: {
        Row: {
          id: number
          created_at: string
          full_name: string
          interests: string
          success_rate: number
          phone: string | null
          email: string | null
          profile_url: string
          post_text: string
          post_summary: string
          message_content: string
        }
        Insert: {
          id?: number
          created_at?: string
          full_name: string
          interests: string
          success_rate: number
          phone?: string | null
          email?: string | null
          profile_url: string
          post_text: string
          post_summary: string
          message_content: string
        }
        Update: {
          id?: number
          created_at?: string
          full_name?: string
          interests?: string
          success_rate?: number
          phone?: string | null
          email?: string | null
          profile_url?: string
          post_text?: string
          post_summary?: string
          message_content?: string
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}